- Changed enemy movement to move in a straight line to target rather than at constant velocities. - Moved level content out of Game class and into new Level class with factory method to hold more content and logic type data.

+++ Object.cpp
@@ -4,7 +4,7 @@
 
 #pragma region Constructor
 
-Object::Object(int x, int y, std::string texturePath, ElementType type)
+Object::Object(double x, double y, std::string texturePath, ElementType type)
 {
 	this->x = x;
 	this->y = y;
